requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
attrs==25.1.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
detectron2==0.6
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
fvcore==0.1.5.post20221221
grpcio==1.70.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imantics==0.1.12
iopath==0.1.9
Jinja2==3.1.5
json5==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lxml==5.3.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.20.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.1.1
protobuf==5.29.3
pycocotools==2.0.8
PyMatting==1.1.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
rembg==2.0.62
requests==2.32.3
rpds-py==0.22.3
scikit-image==0.25.1
scipy==1.15.2
six==1.17.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
tifffile==2025.1.10
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
xmljson==0.2.1
yacs==0.1.8
